+++ src/sentry_scope.c
@@ -312,116 +312,6 @@ sentry__scope_get_span_or_transaction(void)
 }
 #endif
 
-static int
-cmp_breadcrumb(sentry_value_t a, sentry_value_t b, bool *error)
-{
-    sentry_value_t timestamp_a = sentry_value_get_by_key(a, "timestamp");
-    sentry_value_t timestamp_b = sentry_value_get_by_key(b, "timestamp");
-    if (sentry_value_is_null(timestamp_a)) {
-        *error = true;
-        return -1;
-    }
-    if (sentry_value_is_null(timestamp_b)) {
-        *error = true;
-        return 1;
-    }
-
-    return strcmp(sentry_value_as_string(timestamp_a),
-        sentry_value_as_string(timestamp_b));
-}
-
-static bool
-append_breadcrumb(sentry_value_t target, sentry_value_t source, size_t index)
-{
-    int rv = sentry_value_append(
-        target, sentry_value_get_by_index_owned(source, index));
-    if (rv != 0) {
-        SENTRY_ERROR("Failed to merge breadcrumbs");
-        sentry_value_decref(target);
-        return false;
-    }
-    return true;
-}
-
-static sentry_value_t
-merge_breadcrumbs(sentry_value_t list_a, sentry_value_t list_b, size_t max)
-{
-    size_t len_a = sentry_value_get_type(list_a) == SENTRY_VALUE_TYPE_LIST
-        ? sentry_value_get_length(list_a)
-        : 0;
-    size_t len_b = sentry_value_get_type(list_b) == SENTRY_VALUE_TYPE_LIST
-        ? sentry_value_get_length(list_b)
-        : 0;
-
-    if (len_a == 0 && len_b == 0) {
-        return sentry_value_new_null();
-    } else if (len_a == 0) {
-        sentry_value_incref(list_b);
-        return list_b;
-    } else if (len_b == 0) {
-        sentry_value_incref(list_a);
-        return list_a;
-    }
-
-    bool error = false;
-    size_t idx_a = 0;
-    size_t idx_b = 0;
-    size_t total = len_a + len_b;
-    size_t skip = total > max ? total - max : 0;
-    sentry_value_t result = sentry__value_new_list_with_size(total - skip);
-
-    // skip oldest breadcrumbs to fit max
-    while (idx_a < len_a && idx_b < len_b && idx_a + idx_b < skip) {
-        sentry_value_t item_a = sentry_value_get_by_index(list_a, idx_a);
-        sentry_value_t item_b = sentry_value_get_by_index(list_b, idx_b);
-
-        if (cmp_breadcrumb(item_a, item_b, &error) <= 0) {
-            idx_a++;
-        } else {
-            idx_b++;
-        }
-    }
-    while (idx_a < len_a && idx_a + idx_b < skip) {
-        idx_a++;
-    }
-    while (idx_b < len_b && idx_a + idx_b < skip) {
-        idx_b++;
-    }
-
-    // merge the remaining breadcrumbs in timestamp order
-    while (idx_a < len_a && idx_b < len_b) {
-        sentry_value_t item_a = sentry_value_get_by_index(list_a, idx_a);
-        sentry_value_t item_b = sentry_value_get_by_index(list_b, idx_b);
-
-        if (cmp_breadcrumb(item_a, item_b, &error) <= 0) {
-            if (!append_breadcrumb(result, list_a, idx_a++)) {
-                return sentry_value_new_null();
-            }
-        } else {
-            if (!append_breadcrumb(result, list_b, idx_b++)) {
-                return sentry_value_new_null();
-            }
-        }
-    }
-    while (idx_a < len_a) {
-        if (!append_breadcrumb(result, list_a, idx_a++)) {
-            return sentry_value_new_null();
-        }
-    }
-    while (idx_b < len_b) {
-        if (!append_breadcrumb(result, list_b, idx_b++)) {
-            return sentry_value_new_null();
-        }
-    }
-
-    if (error) {
-        SENTRY_WARN("Detected missing timestamps while merging breadcrumbs. "
-                    "This may lead to unexpected results.");
-    }
-
-    return result;
-}
-
 void
 sentry__scope_apply_to_event(const sentry_scope_t *scope,
     const sentry_options_t *options, sentry_value_t event,
@@ -522,8 +412,8 @@ sentry__scope_apply_to_event(const sentry_scope_t *scope,
         sentry_value_t scope_breadcrumbs
             = sentry__ringbuffer_to_list(scope->breadcrumbs);
         sentry_value_set_by_key(event, "breadcrumbs",
-            merge_breadcrumbs(event_breadcrumbs, scope_breadcrumbs,
-                options->max_breadcrumbs));
+            sentry__value_merge_breadcrumbs(event_breadcrumbs,
+                scope_breadcrumbs, options->max_breadcrumbs));
         sentry_value_decref(scope_breadcrumbs);
     }
 
@@ -774,3 +664,87 @@ sentry_scope_attach_bytesw_n(sentry_scope_t *scope, const char *buf,
         ATTACHMENT, NULL);
 }
 #endif
+
+void
+sentry__scope_apply_attributes(const sentry_scope_t *scope,
+    sentry_value_t telemetry, sentry_value_t attributes)
+{
+    sentry__value_merge_objects(attributes, scope->attributes);
+
+    sentry_value_t trace_id = sentry_value_get_by_key(
+        sentry_value_get_by_key(scope->propagation_context, "trace"),
+        "trace_id");
+    sentry_value_incref(trace_id);
+    sentry_value_set_by_key(telemetry, "trace_id", trace_id);
+
+    sentry_value_t parent_span_id = sentry_value_new_object();
+    sentry_value_t scoped_span_trace_id = sentry_value_new_null();
+    if (scope->transaction_object) {
+        sentry_value_t span_id = sentry_value_get_by_key(
+            scope->transaction_object->inner, "span_id");
+        sentry_value_incref(span_id);
+        sentry_value_set_by_key(parent_span_id, "value", span_id);
+        scoped_span_trace_id = sentry_value_get_by_key(
+            scope->transaction_object->inner, "trace_id");
+        sentry_value_incref(scoped_span_trace_id);
+    } else if (scope->span) {
+        sentry_value_t span_id
+            = sentry_value_get_by_key(scope->span->inner, "span_id");
+        sentry_value_incref(span_id);
+        sentry_value_set_by_key(parent_span_id, "value", span_id);
+        scoped_span_trace_id
+            = sentry_value_get_by_key(scope->span->inner, "trace_id");
+        sentry_value_incref(scoped_span_trace_id);
+    }
+    sentry_value_set_by_key(
+        parent_span_id, "type", sentry_value_new_string("string"));
+    if (scope->transaction_object || scope->span) {
+        sentry_value_set_by_key(
+            attributes, "sentry.trace.parent_span_id", parent_span_id);
+        sentry_value_set_by_key(telemetry, "trace_id", scoped_span_trace_id);
+    } else {
+        sentry_value_decref(parent_span_id);
+        sentry_value_decref(scoped_span_trace_id);
+    }
+
+    if (!sentry_value_is_null(scope->user)) {
+        sentry_value_t user_id = sentry_value_get_by_key(scope->user, "id");
+        if (!sentry_value_is_null(user_id)) {
+            sentry_value_incref(user_id);
+            sentry__value_add_attribute(
+                attributes, user_id, "string", "user.id");
+        }
+
+        sentry_value_t user_username
+            = sentry_value_get_by_key(scope->user, "username");
+        if (!sentry_value_is_null(user_username)) {
+            sentry_value_incref(user_username);
+            sentry__value_add_attribute(
+                attributes, user_username, "string", "user.name");
+        }
+
+        sentry_value_t user_email
+            = sentry_value_get_by_key(scope->user, "email");
+        if (!sentry_value_is_null(user_email)) {
+            sentry_value_incref(user_email);
+            sentry__value_add_attribute(
+                attributes, user_email, "string", "user.email");
+        }
+    }
+    sentry_value_t os_context = sentry_value_get_by_key(scope->contexts, "os");
+    if (!sentry_value_is_null(os_context)) {
+        sentry_value_t os_name = sentry_value_get_by_key(os_context, "name");
+        sentry_value_t os_version
+            = sentry_value_get_by_key(os_context, "version");
+        if (!sentry_value_is_null(os_name)) {
+            sentry_value_incref(os_name);
+            sentry__value_add_attribute(
+                attributes, os_name, "string", "os.name");
+        }
+        if (!sentry_value_is_null(os_version)) {
+            sentry_value_incref(os_version);
+            sentry__value_add_attribute(
+                attributes, os_version, "string", "os.version");
+        }
+    }
+}